Finalizacion del Guardar la cotizacion

+++ MandaditosExpress/Models/Cotizacion.cs
@@ -42,7 +42,7 @@ namespace MandaditosExpress.Models
         public bool EsEspecial { get; set; }
 
         [Required]
-        public float MontoTotal { get; set; }
+        public decimal MontoTotal { get; set; }
 
         public int ClienteId { get; set; }
 
@@ -50,7 +50,7 @@ namespace MandaditosExpress.Models
         [Display(Name = "Tipo de Servicio")]
         public int TipoDeServicioId { get; set; }
 
-        public float MontoDeDinero { get; set; }
+        public decimal MontoDeDinero { get; set; }
 
         public int GestionBancariaId { get; set; }
